Replaced Jan Nijtman's Dockerfile with a new one that does a 2-stage build.  The first stage runs atop Alpine Linux instead of Fedora, reducing the initial build from ~635 MiB to about 16.
Rather than stop there, I then made it multi-stage, copying two key
static binaries — Fossil and Busybox — over from the first stage into a
fresh-from-scratch container and set it up to run the former jailed away
from the latter.

The result is under 9 MiB, and it's as secure as one can hope, given
that it starts up in "PUBLIC" mode.  The new build doesn't have all the
extra features turned on that the old one did, but it seems right to
build the container with Fossil in its default configuration. If you
want something else, copy the Dockerfile, hack it, and make it do what
you want instead.

Having done all this, I replaced the one-off Dockerfile inline in section
5.0 of the build doc with a reference to this new Dockerfile and rewrote
the section in terms of the new capabilities.

Finally, this lets us brag on how small the container can be, as
compared to the Gitlab-CE container.  Before, we were comparing a
standalone binary to the container, which wan't entirely fair.  (The
desire to produce such a container was the spark that kicked this
project off.)

FossilOrigin-Name: 77d603c6a194a8941a028a5d43f38122b86114ca6525a5dd2658f3483c99a7a8