The container now uses BusyBox only in the build and setup stages, leaving just the static Fossil binary in the final stage, plus absolute necessities like a /tmp directory.
This removes the justification for the custom BusyBox configuration,
which then means we can use Alpine's busybox-static package in the
second stage, saving a bunch of network I/O and build time.

That in turn means we no longer have any justification for jailing the
Fossil binary, since there's nothing extra left inside the container for
it to play with.  Doing this required bumping the Dockerfile syntax back
up from 1.0 to 1.3 to get the "COPY --chmod" feature; tested it in
Podman, which has had it
[https://github.com/containers/buildah/issues/2961 | for two years now].

Doing all of this simplifies the Dockerfile and its documentation
considerably.  As a bonus, it builds quicker, and it's nearly a meg
lighter in compressed image form.  Especially for the case of using the
container as a static "fossil" binary builder, this is nothing but win.

FossilOrigin-Name: 79ac06a5407ecfe0bb5446241cd61135375f5a5b158562438fc2d8cd4554182c